+++ activerecord/CHANGELOG.md
@@ -1,3 +1,12 @@
+*   Add bigint primary key support for MySQL.
+
+    Example:
+
+        create_table :foos, id: :bigint do |t|
+        end
+
+    *Ryuta Kamizono*
+
 *   Support for any type primary key.
 
     Fixes #14194.